refactor(fetcher): move FetcherRequest interface to fetchExchange.ts
- Move FetcherRequest interface from fetcher.ts to fetchExchange.ts
- Update import statements in fetcher.ts and mergeRequest.ts
- This change prepares for implementing a new RequestFactory class
